# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/event/create": {
            "post": {
                "description": "Хендлер для создания нового мероприятия",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Создание мероприятия",
                "parameters": [
                    {
                        "description": "Мероприятие",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/event/delete": {
            "delete": {
                "description": "Хендлер для удаления мероприятия по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Удалить мероприятие",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID мероприятия",
                        "name": "event_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/event/join": {
            "post": {
                "description": "Хендлер для записи пользователя на мероприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Записаться на мероприятие",
                "parameters": [
                    {
                        "description": "ID мероприятия",
                        "name": "event_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.EventIDRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/event/quit": {
            "post": {
                "description": "Хендлер для отмены записи пользователя на мероприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Отказаться от участия в мероприятии",
                "parameters": [
                    {
                        "description": "ID мероприятия",
                        "name": "event_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.EventIDRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/event/update": {
            "put": {
                "description": "Хендлер для обновления данных мероприятия",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Обновить информацию о мероприятии",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID мероприятия",
                        "name": "event_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateEventInput"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/organizations/join": {
            "post": {
                "description": "Позволяет пользователю присоединиться к организации по заданным ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizations"
                ],
                "summary": "Присоединение пользователя к организации",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID организации",
                        "name": "organization_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/organizations/quit": {
            "post": {
                "description": "Позволяет пользователю выйти из организации. Если пользователь является основателем организации, она будет удалена.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizations"
                ],
                "summary": "Выход пользователя из организации",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID организации",
                        "name": "organization_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "handlers.EventIDRequest": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                }
            }
        },
        "models.Event": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateEventInput": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}